NYC-Taxi-Mobility-Insights/requirements.txt
Flask
pandas
numpy
mysql-connector-python
